feat: Add Terraform configuration for AWS S3 bucket
This commit adds the Terraform configuration files for creating an AWS S3 bucket. The `main.tf` file defines the AWS provider and creates the S3 bucket resource with the specified bucket name. The `outputs.tf` file defines the output variables for the bucket name and ARN. The `terraform.tfvars` file contains the default values for the AWS region and bucket name variables. The `variables.tf` file defines the input variables for the AWS region and bucket name. These configurations enable the creation of an S3 bucket with the desired settings.